unittest: create test util dir; move output capture there
Want to use the OutputCapture in another unit test. There will probably be
more such test helpers over time that need to be imported from different
unit tests.
