+++ tests/test_utils/__init__.py
@@ -1,0 +1,1 @@
+from .output_capture import OutputCapture
